requirements.txt
asgiref==3.8.1
Django==5.1.4
django-unfold==0.43.0
djangorestframework==3.15.2
drf-yasg==1.21.8
environs==11.2.1
inflection==0.5.1
marshmallow==3.23.2
packaging==24.2
pillow==11.0.0
pip-chill==1.0.3
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.3
uritemplate==4.1.1
wheel==0.44.0
